feat(policies): move folder sidebar from editor to policies list page
## Changes
- Remove folder sidebar (FolderTree, file list, CreateFolderModal) from
  PolicyEditorPage.tsx and all related state, hooks, and handlers
- Add folder sidebar as a document browser to PolicyManager.tsx with a
  "Documents" toggle button in the controls row
- Sidebar renders on the left with FolderTree navigation and a scrollable
  file list (read-only, no link action since no policy context)

## Benefits
- Users can browse folders/files alongside the policy list without needing
  to open a specific policy
- Editor page is simplified — folder browsing belongs at the list level
